+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Removed --highmem CLI argument
 - Removed purple highlight for new mods from conflict manager
 - Added --exclude-facegens CLI argument to skip patching facegen meshes
+- Fixed PBR shader not removing Facegen_Detail_Map flag if exists
 
 ## [1.1.4] - 2026-06-24